Initialize the project. * Remove the onboarding code. * Add structure to the project.
