+++ lib/constants.dart
@@ -1,0 +1,3 @@
+class Constants {
+  static const appName = 'Recipe';
+}
